requirements.txt
google-cloud-documentai>=2.20.0
google-cloud-storage>=2.10.0
python-dotenv>=1.0.0
google-api-python-client>=2.90.0
google-cloud-discoveryengine>=0.1.0
google-generativeai>=0.1.0
